Remove the splash screen and fix the loading issue
