+++ forge.config.js
@@ -137,11 +137,6 @@ const config = {
                             js: './src/renderer/renderer.js',
                             name: 'main_window',
                         },
-                        {
-                            html: './src/renderer/splash.html',
-                            js: './src/renderer/splash_renderer.js',
-                            name: 'splash_screen',
-                        },
                     ],
                 },
                 devContentSecurityPolicy: '',
