+++ src/common/locales/en/translation.json
@@ -50,7 +50,6 @@
       "paste": "Paste"
     }
   },
-  "loading": "Loading...",
   "outputs": {
     "largeImage": {
       "imageNotAvailable": "Image not available."
