+++ src/renderer/contexts/BackendContext.tsx
@@ -55,6 +55,7 @@ interface BackendContextState {
     backendDownRef: React.MutableRefObject<boolean>;
     restart: () => Promise<void>;
     connectionState: 'connecting' | 'connected' | 'failed';
+    backendReady: boolean;
 }
 
 export const BackendContext = createContext<Readonly<BackendContextState>>(
@@ -229,6 +230,7 @@ const useNodes = (backend: Backend, backendDownRef: Readonly<MutableRefObject<bo
         scope,
         refreshNodes,
         connectionState,
+        backendReady,
     };
 };
 
@@ -291,10 +293,8 @@ export const BackendProvider = memo(
         const restartPromiseRef = useRef<Promise<void>>();
         const needsNewRestartRef = useRef(false);
 
-        const { nodesInfo, schemaInputs, scope, refreshNodes, connectionState } = useNodes(
-            backend,
-            backendDownRef
-        );
+        const { nodesInfo, schemaInputs, scope, refreshNodes, connectionState, backendReady } =
+            useNodes(backend, backendDownRef);
         const { featureStates, refreshFeatureStates } = useFeatureStates(backend);
 
         const featureStatesMaps = useMemo((): ReadonlyMap<FeatureId, FeatureState> => {
@@ -373,6 +373,7 @@ export const BackendProvider = memo(
             backendDownRef,
             restart,
             connectionState,
+            backendReady,
         });
 
         return <BackendContext.Provider value={value}>{children}</BackendContext.Provider>;
