--- src/renderer/splash.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-  <meta charset="UTF-8">
-  <title>chaiNNer</title>
-
-</head>
-
-<body>
-  <div id="root">
-
-  </div>
-</body>
-
-</html>
