--- src/renderer/splashTheme.ts
@@ -1,12 +1,0 @@
-import { extendTheme } from '@chakra-ui/react';
-
-export const theme = extendTheme({
-    styles: {
-        global: {
-            // styles for the `body`
-            body: {
-                bg: 'none',
-            },
-        },
-    },
-});
