--- src/renderer/splash_renderer.js
@@ -1,2 +1,0 @@
-import './global.scss';
-import './splash';
